requirements.txt
allure-pytest
pytest
pytest-bdd
pytest-cov
pytest-html
pytest-xdist
requests
pyyaml
selenium
webdriver-manager
black
flake8